[Refactor] [FE-49] FAQ 컴포넌트 분리
